Merging develop with main

+++ src/main/java/org/individualproject/business/ExcursionService.java
@@ -1,30 +1,35 @@
 package org.individualproject.business;
 
+import lombok.AllArgsConstructor;
 import org.individualproject.business.converter.ExcursionConverter;
+import org.individualproject.business.converter.UserConverter;
 import org.individualproject.business.exception.InvalidExcursionDataException;
+import org.individualproject.business.exception.UnauthorizedDataAccessException;
+import org.individualproject.configuration.security.token.AccessToken;
 import org.individualproject.domain.CreateExcursionRequest;
 import org.individualproject.domain.Excursion;
 import org.individualproject.domain.UpdateExcursionRequest;
+import org.individualproject.domain.User;
+import org.individualproject.domain.enums.UserRole;
+import org.individualproject.persistence.BookingRepository;
 import org.individualproject.persistence.ExcursionRepository;
+import org.individualproject.persistence.PaymentDetailsRepository;
+import org.individualproject.persistence.entity.BookingEntity;
 import org.individualproject.persistence.entity.ExcursionEntity;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
+import org.individualproject.persistence.entity.UserEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.util.StringUtils;
+import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
+@AllArgsConstructor
 public class ExcursionService {
     private ExcursionRepository excursionRepository;
-    @Autowired
-    public ExcursionService(ExcursionRepository exRepository){
-        this.excursionRepository = exRepository;
-    }
+    private BookingRepository bookingRepository;
+    private PaymentDetailsRepository paymentDetailsRepository;
+    private AccessToken requestAccessToken;
     public List<Excursion> getExcursions() {
         List<ExcursionEntity> excursionEntities = excursionRepository.findAll();
         return ExcursionConverter.mapToDomainList(excursionEntities);
@@ -35,47 +40,100 @@ public class ExcursionService {
     }
 
     public Excursion createExcursion(CreateExcursionRequest request){
-        if (request.getName() == null || request.getDestinations() == null || request.getStartDate() == null ||
+        if (request.getName() == null || request.getDestinations() == null || request.getDescription() == null || request.getStartDate() == null ||
                 request.getEndDate() == null || request.getTravelAgency() == null || request.getPrice() < 0 ||
                 request.getNumberOfAvaliableSpaces() < 0) {
             throw new InvalidExcursionDataException("Invalid input data");
         }
+
+        if (!requestAccessToken.hasRole(UserRole.TRAVELAGENCY.name())) {
+            throw new UnauthorizedDataAccessException("Only travel agencies can list excursions!");
+        }
+
+        UserEntity userEntity = UserConverter.convertToEntity(request.getTravelAgency());
         ExcursionEntity newExcursion = ExcursionEntity.builder()
                 .name(request.getName())
                 .destinations(String.join(",", request.getDestinations()))
+                .description(request.getDescription())
                 .startDate(request.getStartDate())
                 .endDate(request.getEndDate())
-                .travelAgency(request.getTravelAgency())
+                .travelAgency(userEntity)
                 .price(request.getPrice())
                 .numberOfAvaliableSpaces(request.getNumberOfAvaliableSpaces())
+                .numberOfSpacesLeft(request.getNumberOfAvaliableSpaces())
                 .build();
 
         ExcursionEntity excursionEntity = excursionRepository.save(newExcursion);
         return ExcursionConverter.mapToDomain(excursionEntity);
     }
 
+    @Transactional
     public boolean deleteExcursion(Long id) {
-        try {
+        Optional<ExcursionEntity> optionalExcursion = excursionRepository.findById(id);
+
+        if (optionalExcursion.isPresent()) {
+            ExcursionEntity excursion = optionalExcursion.get();
+            if (!requestAccessToken.hasRole(UserRole.TRAVELAGENCY.name())) {
+                throw new UnauthorizedDataAccessException("ONLY_TRAVEL_AGENCIES_ALLOWED");
+            }
+
+            if (!requestAccessToken.getUserID().equals(excursion.getTravelAgency().getId())) {
+                throw new UnauthorizedDataAccessException("EXCURSION_NOT_OWNED_BY_LOGGED_IN_USER");
+            }
+
+            List<BookingEntity> bookings = bookingRepository.findByExcursion(excursion);
+
+            bookingRepository.deleteByExcursion(excursion);
+            for (BookingEntity booking : bookings) {
+                paymentDetailsRepository.deleteById(booking.getBankingDetails().getId());
+            }
             excursionRepository.deleteById(id);
             return true;
-        } catch (EmptyResultDataAccessException e) {
+        } else {
             return false;
         }
     }
 
     public boolean updateExcursion(UpdateExcursionRequest request) {
         Optional<ExcursionEntity> optionalExcursion = excursionRepository.findById(request.getId());
+
+        if (request.getName() == null || request.getDestinations() == null || request.getDescription() == null  || request.getStartDate() == null ||
+                request.getEndDate() == null || request.getPrice() < 0 ||
+                request.getNumberOfAvaliableSpaces() < 0) {
+            throw new InvalidExcursionDataException("Invalid input data");
+        }
         if (optionalExcursion.isPresent()) {
             ExcursionEntity existingExcursion = optionalExcursion.get();
+
+            if (!requestAccessToken.hasRole(UserRole.TRAVELAGENCY.name())) {
+                throw new UnauthorizedDataAccessException("ONLY_TRAVEL_AGENCIES_ALLOWED");
+            }
+
+            if (!requestAccessToken.getUserID().equals(existingExcursion.getTravelAgency().getId())) {
+                throw new UnauthorizedDataAccessException("EXCURSION_NOT_OWNED_BY_LOGGED_IN_USER");
+            }
+            int bookedSpaces = existingExcursion.getNumberOfAvaliableSpaces() - existingExcursion.getNumberOfSpacesLeft();
+
+            int newAvailableSpaces = request.getNumberOfAvaliableSpaces();
+            int newSpacesLeft = existingExcursion.getNumberOfSpacesLeft();
+
+            if (newAvailableSpaces < newSpacesLeft) {
+                newSpacesLeft = newAvailableSpaces - bookedSpaces;
+                if (newSpacesLeft < 0) {
+                    newSpacesLeft = 0;
+                }
+            }
+
             existingExcursion.setName(request.getName());
             List<String> destinations = request.getDestinations();
             String destinationsString = String.join(",", destinations);
             existingExcursion.setDestinations(destinationsString);
+            existingExcursion.setDescription(request.getDescription());
             existingExcursion.setStartDate(request.getStartDate());
             existingExcursion.setEndDate(request.getEndDate());
-            existingExcursion.setTravelAgency(request.getTravelAgency());
             existingExcursion.setPrice(request.getPrice());
-            existingExcursion.setNumberOfAvaliableSpaces(request.getNumberOfAvaliableSpaces());
+            existingExcursion.setNumberOfAvaliableSpaces(newAvailableSpaces);
+            existingExcursion.setNumberOfSpacesLeft(newSpacesLeft);
             excursionRepository.save(existingExcursion);
             return true;
         } else {
@@ -88,48 +146,61 @@ public class ExcursionService {
         return excursionEntity.map(ExcursionConverter::mapToDomain);
     }
 
-    public List<Excursion> findExcursionsByName(String name){
-        List<ExcursionEntity> excursionEntities = excursionRepository.findByNameContainingIgnoreCase(name);
-        return ExcursionConverter.mapToDomainList(excursionEntities);
-    }
 
-    public List<Excursion> findExcursionsByPriceRange(String priceRange){
-        List<ExcursionEntity> excursionEntities;
-        if (priceRange.startsWith("<") || priceRange.startsWith(">")) {
-            double price = Double.parseDouble(priceRange.substring(1)); // Remove the ">" or "<" symbol
-            if (priceRange.startsWith(">")) {
-                excursionEntities = excursionRepository.findByPriceGreaterThan(price);
-            } else {
-                excursionEntities = excursionRepository.findByPriceLessThan(price);
-            }
-        } else {
-            String[] prices = priceRange.split("-");
-            double minPrice = Double.parseDouble(prices[0]);
-            double maxPrice = Double.parseDouble(prices[1]);
-            excursionEntities = excursionRepository.findByPriceRange(minPrice, maxPrice);
+    public void bookSpaces(Long id, int spacesBooked){
+        int updatedRows = excursionRepository.decrementSpacesLeft(id, spacesBooked);
+        if(updatedRows == 0){
+            throw new IllegalStateException("Not enough spaces left for this excursion!");
         }
+    }
+
+    public List<Excursion> getExcursionsByTravelAgency(User travelAgency) {
+        UserEntity userEntity = UserConverter.convertToEntity(travelAgency);
+        List<ExcursionEntity> excursionEntities = excursionRepository.findByTravelAgency(userEntity);
         return ExcursionConverter.mapToDomainList(excursionEntities);
     }
 
-    public List<Excursion> searchExcursionsByNameAndPriceRange(String name, String priceRange) {
-        List<ExcursionEntity> excursionEntities = new ArrayList<>();
-        if (priceRange != null && !priceRange.isEmpty()) {
-            if (priceRange.startsWith(">") || priceRange.endsWith(">")) {
-                double price = Double.parseDouble(priceRange.substring(1)); // Remove the ">" symbol
-                excursionEntities = excursionRepository.findByPriceGreaterThan(price);
-            } else if (priceRange.startsWith("<") || priceRange.endsWith("<")) {
-                double price = Double.parseDouble(priceRange.substring(1)); // Remove the "<" symbol
-                excursionEntities = excursionRepository.findByPriceLessThan(price);
-            } else {
-                String[] prices = priceRange.split("-");
-                double minPrice = Double.parseDouble(prices[0]);
-                double maxPrice = Double.parseDouble(prices[1]);
-                excursionEntities = excursionRepository.findByPriceRange(minPrice, maxPrice);
-            }
-        } else {
-            excursionEntities = excursionRepository.findByNameContainingIgnoreCase(name);
+    public List<Excursion> searchExcursionsByNameAndTravelAgency(String searchTerm) {
+        List<ExcursionEntity> excursionEntities = excursionRepository.findByNameContainingIgnoreCaseOrTravelAgency_FirstNameContainingIgnoreCaseOrTravelAgency_LastNameContainingIgnoreCase(searchTerm, searchTerm, searchTerm);
+        return ExcursionConverter.mapToDomainList(excursionEntities);
+    }
+
+    public List<Excursion> searchExcursions(String searchTerm, double minPrice, double maxPrice) {
+        if ((searchTerm == null || searchTerm.isEmpty()) && (minPrice == 0 && maxPrice == Double.MAX_VALUE)) {
+            return ExcursionConverter.mapToDomainList(excursionRepository.findAll());
         }
-        return ExcursionConverter.mapToDomainList(excursionEntities);
+
+        if (searchTerm == null) {
+            searchTerm = "";
+        }
+
+        if (searchTerm.isEmpty() && (minPrice > 0 || maxPrice < Double.MAX_VALUE)) {
+            return ExcursionConverter.mapToDomainList(
+                    excursionRepository.findByPriceBetween(minPrice, maxPrice)
+            );
+        } else if (searchTerm.isEmpty()) {
+            return ExcursionConverter.mapToDomainList(excursionRepository.findAll());
+        } else if (minPrice == 0 && maxPrice == Double.MAX_VALUE) {
+            return ExcursionConverter.mapToDomainList(
+                    excursionRepository.findByNameContainingIgnoreCaseOrTravelAgency_FirstNameContainingIgnoreCaseOrTravelAgency_LastNameContainingIgnoreCase(
+                            searchTerm, searchTerm, searchTerm
+                    )
+            );
+        } else {
+            return ExcursionConverter.mapToDomainList(
+                    excursionRepository.findByNameContainingIgnoreCaseAndPriceBetweenOrTravelAgency_FirstNameContainingIgnoreCaseAndPriceBetweenOrTravelAgency_LastNameContainingIgnoreCaseAndPriceBetween(
+                            searchTerm, minPrice, maxPrice,
+                            searchTerm, minPrice, maxPrice,
+                            searchTerm, minPrice, maxPrice
+                    )
+            );
+        }
     }
 
+    public void updateExcursionPresentationFile(Long excursionId, String fileName) {
+        ExcursionEntity excursion = excursionRepository.findById(excursionId)
+                .orElseThrow(() -> new RuntimeException("Excursion not found with id: " + excursionId));
+        excursion.setFileName(fileName);
+        excursionRepository.save(excursion);
+    }
 }
