+++ src/main/java/org/individualproject/business/PaymentDetailsService.java
@@ -1,35 +1,55 @@
 package org.individualproject.business;
 
+import lombok.AllArgsConstructor;
 import org.individualproject.business.converter.PaymentDetailsConverter;
 import org.individualproject.business.converter.UserConverter;
+import org.individualproject.business.exception.InvalidExcursionDataException;
+import org.individualproject.business.exception.UnauthorizedDataAccessException;
+import org.individualproject.configuration.security.token.AccessToken;
 import org.individualproject.domain.*;
+import org.individualproject.domain.enums.UserRole;
 import org.individualproject.persistence.PaymentDetailsRepository;
 import org.individualproject.persistence.entity.PaymentDetailsEntity;
 import org.individualproject.persistence.entity.UserEntity;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
 @Service
+@AllArgsConstructor
 public class PaymentDetailsService {
+    private static final String UNAUTHORIZED_ACCESS = "UNAUTHORIZED_ACCESS";
+
     private PaymentDetailsRepository paymentDetailsRepository;
-    @Autowired
-    public PaymentDetailsService(PaymentDetailsRepository paymentDRepository){
-        this.paymentDetailsRepository = paymentDRepository;
-    }
+    private AccessToken requestAccessToken;
     public List<PaymentDetails> getAllPaymentDetails() {
         List<PaymentDetailsEntity> paymentDetailsEntities = paymentDetailsRepository.findAll();
         return PaymentDetailsConverter.mapToDomainList(paymentDetailsEntities);
     }
     public Optional<PaymentDetails> getPaymentDetails(Long id) {
-        Optional<PaymentDetailsEntity> paymentDetailsEntities = paymentDetailsRepository.findById(id);
-        return paymentDetailsEntities.map(PaymentDetailsConverter::mapToDomain);
+
+        Optional<PaymentDetailsEntity> paymentDetailsOptional = paymentDetailsRepository.findById(id);
+
+        if (paymentDetailsOptional.isPresent()) {
+            PaymentDetailsEntity paymentDetailsEntity = paymentDetailsOptional.get();
+
+            if (!requestAccessToken.hasRole(UserRole.TRAVELAGENCY.name()) &&
+                    !requestAccessToken.getUserID().equals(paymentDetailsEntity.getUser().getId())) {
+                throw new UnauthorizedDataAccessException(UNAUTHORIZED_ACCESS);
+            }
+            return Optional.of(PaymentDetailsConverter.mapToDomain(paymentDetailsEntity));
+            //return paymentDetailsOptional.map(PaymentDetailsConverter::mapToDomain);
+        } else {
+            return Optional.empty();
+        }
     }
 
     public PaymentDetails createPaymentDetails(CreatePaymentDetailsRequest request){
+        if (request.getCardHolderName() == null || request.getCardNumber() == null || request.getCvv() == null ||
+                request.getUser() == null || request.getExpirationDate() == null) {
+            throw new InvalidExcursionDataException("Invalid input data");
+        }
         UserEntity userEntity = UserConverter.convertToEntity(request.getUser());
         PaymentDetailsEntity newPaymentDetails = PaymentDetailsEntity.builder()
                 .user(userEntity)
@@ -44,19 +64,37 @@ public class PaymentDetailsService {
     }
 
     public boolean deletePaymentDetails(Long id) {
-        try {
+        Optional<PaymentDetailsEntity> paymentDetailsOptional = paymentDetailsRepository.findById(id);
+        if (paymentDetailsOptional.isPresent()) {
+            PaymentDetailsEntity paymentDetailsEntity = paymentDetailsOptional.get();
+
+
+            if (!requestAccessToken.hasRole(UserRole.TRAVELAGENCY.name()) &&
+                    !requestAccessToken.getUserID().equals(paymentDetailsEntity.getUser().getId())) {
+                throw new UnauthorizedDataAccessException(UNAUTHORIZED_ACCESS);
+            }
+
             paymentDetailsRepository.deleteById(id);
             return true;
-        } catch (EmptyResultDataAccessException e) {
+        } else {
             return false;
         }
     }
 
     public boolean updatePaymentDetails(UpdatePaymentDetailsRequest request) {
+        if (request.getCardHolderName() == null || request.getCardNumber() == null || request.getCvv() == null ||
+                request.getUser() == null || request.getExpirationDate() == null) {
+            throw new InvalidExcursionDataException("Invalid input data");
+        }
         UserEntity userEntity = UserConverter.convertToEntity(request.getUser());
         Optional<PaymentDetailsEntity> optionalPaymentDetails = paymentDetailsRepository.findById(request.getId());
         if (optionalPaymentDetails.isPresent()) {
             PaymentDetailsEntity existingPaymentDetails = optionalPaymentDetails.get();
+
+            if (!requestAccessToken.hasRole(UserRole.TRAVELAGENCY.name()) &&
+                    !requestAccessToken.getUserID().equals(existingPaymentDetails.getUser().getId())) {
+                throw new UnauthorizedDataAccessException(UNAUTHORIZED_ACCESS);
+            }
             existingPaymentDetails.setId(request.getId());
             existingPaymentDetails.setUser(userEntity);
             existingPaymentDetails.setCvv(request.getCvv());
