+++ src/main/java/org/individualproject/business/UserService.java
@@ -1,33 +1,49 @@
 package org.individualproject.business;
 
-import org.individualproject.business.converter.ExcursionConverter;
+import lombok.AllArgsConstructor;
 import org.individualproject.business.converter.UserConverter;
+import org.individualproject.business.exception.NotFoundException;
+import org.individualproject.business.exception.UnauthorizedDataAccessException;
+import org.individualproject.configuration.security.token.AccessToken;
 import org.individualproject.domain.*;
-import org.individualproject.persistence.ExcursionRepository;
-import org.individualproject.persistence.UserRepository;
-import org.individualproject.persistence.entity.ExcursionEntity;
+import org.individualproject.persistence.*;
+import org.individualproject.persistence.entity.BookingEntity;
 import org.individualproject.persistence.entity.UserEntity;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
+@AllArgsConstructor
+
 public class UserService {
 
     private UserRepository userRepository;
-    @Autowired
-    public UserService(UserRepository uRepository){
-        this.userRepository = uRepository;
-    }
-    public List<User> getUsers() {
+    private ReviewRepository reviewRepository;
+    private BookingRepository bookingRepository;
+    private PaymentDetailsRepository paymentDetailsRepository;
+    private ExcursionRepository excursionRepository;
+    private NotificationsRepository notificationsRepository;
+    private AccessToken requestAccessToken;
+   public List<User> getUsers() {
         List<UserEntity> userEntities = userRepository.findAll();
         return UserConverter.mapToDomainList(userEntities);
     }
     public Optional<User> getUser(Long id) {
         Optional<UserEntity> userEntity = userRepository.findById(id);
+        if (userEntity.isPresent()) {
+            return Optional.of(UserConverter.mapToDomain(userEntity.get()));
+        } else {
+            throw new NotFoundException("User not found");
+        }
+    }
+
+    public Optional<User> getUserByUsername(String username) {
+        Optional<UserEntity> userEntity = userRepository.findByUsername(username);
         return userEntity.map(UserConverter::mapToDomain);
     }
 
@@ -37,26 +53,51 @@ public class UserService {
                 .lastName(request.getLastName())
                 .birthDate(request.getBirthDate())
                 .email(request.getEmail())
-                .password(request.getPassword())
-                .hashedPassword(request.getHashedPassword())
-                .salt(request.getSalt())
+                .username(request.getUsername())
+                .hashedPassword(request.getPassword())
                 .gender(request.getGender())
                 .build();
 
         UserEntity userEntity = userRepository.save(newUser);
-        return UserConverter.mapToDomain(userEntity);
+        if (userEntity.getId() != null) {
+            return UserConverter.mapToDomain(userEntity);
+        } else {
+            return null;
+        }
     }
 
+    @Transactional
     public boolean deleteUser(Long id) {
         try {
-            userRepository.deleteById(id);
-            return true;
+            if (!Objects.equals(requestAccessToken.getUserID(), id)) {
+                throw new UnauthorizedDataAccessException("USER_ID_NOT_FROM_LOGGED_IN_USER");
+            }
+            Optional<UserEntity> user = userRepository.findById(id);
+
+            if(user.isPresent()){
+                UserEntity userEntity = user.get();
+                notificationsRepository.deleteByUserId(userEntity.getId());
+                reviewRepository.deleteByUserWriter(userEntity);
+                reviewRepository.deleteByTravelAgency(userEntity);
+                bookingRepository.deleteByUser(userEntity);
+                paymentDetailsRepository.deleteByUser(userEntity);
+
+                excursionRepository.deleteByTravelAgency(userEntity);
+
+                userRepository.deleteById(id);
+                return true;
+            }
+            return false;
         } catch (EmptyResultDataAccessException e) {
             return false;
         }
     }
 
     public boolean updateUser(UpdateUserRequest request) {
+        if (!requestAccessToken.getUserID().equals(request.getId())) {
+            throw new UnauthorizedDataAccessException("USER_ID_NOT_FROM_LOGGED_IN_USER");
+        }
+
         Optional<UserEntity> optionalUser = userRepository.findById(request.getId());
         if (optionalUser.isPresent()) {
             UserEntity existingUser = optionalUser.get();
@@ -73,12 +114,11 @@ public class UserService {
             if(existingUser.getBirthDate()!= null){
                 existingUser.setBirthDate(request.getBirthDate());
             }
-            userRepository.save(existingUser); // This should not trigger password validation
+            userRepository.save(existingUser);
             return true;
         } else {
             return false;
         }
     }
-
 
 }
